One Front Door/SourceCode/MerchAngular2ReleaseiPhone/src/app/pl-merch/pl-route-item/pl-route-item.component.spec.ts
import {
  beforeEach,
  beforeEachProviders,
  describe,
  expect,
  it,
  inject,
} from '@angular/core/testing';
import { ComponentFixture, TestComponentBuilder } from '@angular/compiler/testing';
import { Component } from '@angular/core';
import { By } from '@angular/platform-browser';
import { PlRouteItemComponent } from './pl-route-item.component';

describe('Component: PlRouteItem', () => {
  let builder: TestComponentBuilder;

  beforeEachProviders(() => [PlRouteItemComponent]);
  beforeEach(inject([TestComponentBuilder], function (tcb: TestComponentBuilder) {
    builder = tcb;
  }));

  it('should inject the component', inject([PlRouteItemComponent],
      (component: PlRouteItemComponent) => {
    expect(component).toBeTruthy();
  }));

  it('should create the component', inject([], () => {
    return builder.createAsync(PlRouteItemComponentTestController)
      .then((fixture: ComponentFixture<any>) => {
        let query = fixture.debugElement.query(By.directive(PlRouteItemComponent));
        expect(query).toBeTruthy();
        expect(query.componentInstance).toBeTruthy();
      });
  }));
});

@Component({
  selector: 'test',
  template: `
    <app-pl-route-item></app-pl-route-item>
  `,
  directives: [PlRouteItemComponent]
})
class PlRouteItemComponentTestController {
}

